Lots of changes. Ready for 1.3 release
Display output for all warranties for a system. Add up the
number of days left to give an accurate count of the time remaining. Fix
basic check for Dell's database being down. Fixed regex to be non-greedy.
Start and end dates for warranty now takes all warranties into account.
Date output is now yyyy-mm-dd because that is more international.